Refactor some more
