+++ app/redux/reducers/index.js
@@ -1,0 +1,7 @@
+// Library Imports
+import { combineReducers } from "redux-immutable";
+
+// Local imports
+import { location } from "./location";
+
+export default rootReducer = combineReducers({ location });
