+++ app/redux/reducers/utils.js
@@ -1,0 +1,2 @@
+const applyFn = (state, fn) => fn(state);
+export const pipe = (fns, state) => state.withMutations(s => fns.reduce(applyFn, s));
